library: initial commit
ioucontext is a coöperative multitasking framework built from liburing [1]
and libucontext [2] (or libc built-in); it is a simplified spiritual
successor to liboco [3].

Architecturally, each system thread has a thread-local reactor that includes
an io_uring instance, a wait queue, and an optional user-data cookie.  The
reactor is initialized just-in-time by `reactor_get` and cleaned up at
thread exit.  The underlying threads are C11, not necessarily pthreads.

During reactor initialization, ioucontext will attempt to pin the underlying
io_ring to one CPU for which a reactor's thread has scheduling affinity.
The intention is to have one io_uring running per thread, each pinned to one
processor.

New tasks can be added to a reactor via `reactor_fiber` and invoking
`reactor_run` will kick off an attempt to run all fibers to completion.
Each fiber owns its own stack, which is cleaned up when the fiber returns.

Though fibers are create with ucontext, the reactor uses siglongjmp to avoid
the syscall overhead of changing signal masks on systems that are POSIX
compliant on this point.

Multiple IO operations are implemented via liburing and take a handle to the
thread-local reactor.  Operations are added either to the io_uring submit
queue, or to a wait queue on overflow.  The queue depth can be controlled by
setting `IOUCONTEXT_QUEUE_DEPTH` in the environment.  Note that some
operations require multiple adjacent submit queue entries to start and WILL
block the entire reactor if the queue depth is set too low.  The results of
an operation are returned to the caller, with error states as negative
return values vs. via `errno`.

Operations MUST NOT be invoked outside of a fiber.

Context-switches only happen during operations, but can be forced via
`io_barrier`, `io_flush`, & `io_yield`.

User-data can be attached to a reactor via `cookie_jar`, including an
optional destructor, and retrieved via `cookie_get`.  The cookie can be
destroyed via `cookie_eat`, but replacing a cookie will not invoke its
destructor.

In order to shut down the reactor from a fiber, all pending operations must
be cleared or canceled.  Of note, io_cancel_any and io_cancel_all do NOT
block to avoid a context-switch into the reactor.

[1] https://github.com/axboe/liburing
[2] https://github.com/kaniini/libucontext
[3] https://github.com/pallas/liboco