iou_port7: initial commit
iou_port7 launches rfc862 echo protocol [1] services on TCP port 12345 and
UDP port 12345 (note: not port 7).  For UDP, the payload of the recieved
datagram is sent back to the peer.  For TCP, the stream ingress is spliced
into a pipe and the pipe is spliced into the stream egress.  The pipe buffer
sizes are adjusted to match the stream.  Compare to port7 [2], which runs on
libevent2 [3].

Several signals are blocked for the process, handled instead by a signalfd
fiber.  The UDP recv and TCP listen file descriptors are registered with
this signal handler via a user-data cookie.  On SIGINT or SIGHUP, the
handler cancels in-progress operations for these FDs, preventing new
connections.  However, existing TCP streams are not canceled and the reactor
will continue to run handler fibers until they shutdown naturally.

This example demonstrates how to launch multiple services, how to shutdown
listening fibers while allowing other fibers to complete normally, and how
to utilize user-data cookies.

[1] https://datatracker.ietf.org/doc/html/rfc862
[2] https://github.com/pallas/port7
[3] https://libevent.org/