reactor: refactor core, allow operations outside of fiber
This change modifies the `reactor_enter_core` to directly process CQEs and
pending TODOs, rather than trying to jump back into `reactor_run`.  Doing so
greatly simplifies the logic AND enables invocations of IO operations
outside of the context of a `reactor_fiber`, with the caveat that the
reactor may still have work to do when the operations completes.

The predicate `reactor_runnable` indicates that there is more work to do and
a call to `reactor_run` will complete it.