fiber: fix stack cleanup
Previously, `fiber_put` ran on the reactor's shared scratch stack.  This
works on platforms that pass the first argument in registers, since those
get saved in the ucontext.  But for ABIs that put all arguments on the
stack, like x86, fibers would clobber each other wrt the bounce.

Instead, slice a small piece of the fiber stack off to use for the bounce
and then switch to the scratch stack in order to `munmap` the fiber stack.