reactor: refactor core, prevent a deadlock
There was some unnecessary hoop jumping in `reactor_enter_core` related to
submitting outstanding SQEs and waiting for CQEs to show up.  Consequently,
if `reactor_cqes` processed the last in-flight CQE and it did not jump out
of the core, `reactor_enter_core` would hang the reactor waiting for a CQE
that would never arrive.  This doesn't happen when the last CQE does jump
away because the way back in is through a direct call and the outer loop
will check `reactor_runnable` before processing todos or io_uring.

This commit moves ALL of the waiting `reactor_cqes` and assumes that it will
only be invoked when `reactor_inflight` is true.  Additionally, the result
is only ever updated if we expect to invoke a todo that will use it, and the
`reactor_todos` method replaces `jump_queue_empty` in several places.

Making todo processing a higher priority was intentional, since defered work
causes new work to be defered as well.