iou_dns: initial commit
This commit adds both iou_dns and async DNS resolution via the c-ares
library [1].

iou_dns takes a series of domain names and network addresses on the command
line.  For each domain name, it spawns a thread that attempts to resolve the
name into an address.  For each network address, it spawns a thread that
attempts to reverse the address into a name.  This resolution happens
asynchronously, with one exception: a TCP `connect` to a non-existant server
can hang the resolver.  Note that this is behavior coming from c-ares, not
iou_context.

Basic name-to-addr looks like
```c
	reactor_t * reactor = reactor_get();

	iou_ares_data_t iou_ares_data;
	ares_channel_t * channel = iou_ares_get(reactor, &iou_ares_data, NULL, 0);

	iou_ares_addr_result_t result;
	iou_ares_addrinfo(iou_ares_data, "example.net", NULL, NULL, &result);
	iou_ares_wait(&result.future);
	if (ARES_SUCCESS != result.status) {
		// result.addrinfo is valid
		ares_freeaddrinfo(result.addrinfo);
	}

	iou_ares_put(&iou_ares_data);
```

Basic addr-to-name looks like
```c
	reactor_t * reactor = reactor_get();

	iou_ares_data_t iou_ares_data;
	ares_channel_t * channel = iou_ares_get(reactor, &iou_ares_data, NULL, 0);

	const struct sockaddr_storage sockaddr;
	socklen_t socklen = sockaddr_parse(&sockaddr, "::1", 0);

	iou_ares_name_result_t result;
	iou_ares_nameinfo(iou_ares_data, (const struct sockaddr *)&sockaddr, socklen, 0, &result);
	iou_ares_wait(&result.future);
	if (ARES_SUCCESS != result.status) {
		// result.node and result.service are valid-ish
		free(result.node);
		free(result.service);
	}

	iou_ares_put(&iou_ares_data);
```

Note that `iou_ares_wait` can take multiple result futures and will wait for
all of them to complete before it returns, e.g.

```c
	iou_ares_wait(&result[0].future, &result[1].future, ... , &result[N].future);
```

[1] https://c-ares.org/