reactor/cqes: ignore spurious wakeup issue
Prior to the rewrite in [e32e783 reactor: refactor core, prevent a
deadlock], the `io_uring_for_each_cqe` was protected by a call to
`io_uring_cq_ready`, which was replaced with a blocking call to
`io_uring_submit_and_wait`.

Afterward, the reactor was observed sometimes entering a state where

	1. `io_uring_submit_and_wait` unblocked, but
	2. `io_uring_for_each_cqe` iterated through zero CQEs, thus
	3. `delta` was zero, violating an assertion.

Removing the assertion does not appear to cause any immediate issues.