operations: add spawn
This method takes a command path, a list of arguments, an optional spawn
attr, & optional landing-zones for stdin/stdout pipes; it spawns a new
process using `posix_spawn`.

If `to_fd` is NULL, STDIN is closed for the process.  Otherwise, a pipe is
hooked up to STDIN and `to_fd` is set to the write end.

If `from_fd` is NULL, STDOUT is connected to /dev/null.  Otherwise, a pipe
is hooked up to STDOUT and `from_fd` is set to the read end.

Because fiber stacks are not mapped across `fork`, the call to `posix_spawn`
happens in a temporary ucontext running on the reactor's scratch stack.  In
addition, the contents of `argv` and `attrs` are copied to this stack.

The interface is an attempt to strike a balance between power and complexity
and may change in the future.