operations: add epoll_ctl
This io_uring operation was left out before but is useful for some coming
changes.  In addition to the basic `iou_epoll_ctl`, there are helpers for

 * iou_poll_add --- add a watch on `fd` if it does not exist
 * iou_poll_del --- delete a watch on `fd` if it does exist
 * iou_poll_mod --- modify a watch on `fd` if it does exist
 * iou_poll_set --- add or modify a watch on `fd`

In particular, `iou_poll_set` first attempts to modify a watch and if the
watch does not exist, adds it.  This is because in a sequence of

	add, mod1, ... modN, del

there will be many more successful `mod`s than `del`s.