fiber: typecheck makecontext
The `reactor_fiber` macro passes arguments directly to `make_context`, which
does not typecheck the arguments against the provided function.  This change
creates an unused variable of type `typeof(function(args...)) *`, which
forces type checking.  This is important because when `setcontext` is called
on a `ucontext_t` where there is an argument mismatch, it will fail wildly.