reactor: promise nonchalance
Impatient promises have an absolute timeout that is not affected by waiting
for SQEs and may be in the past by the time an SQE is available.  This means
that simple non-blocking requests may fail just because they had to wait to
run, or spuriously depending on the scheduler.

In contrast, nonchalant promises run with a time relative to submission and
have a 32k ns grace period.  Rudimentary testing showed this value to
provide a good balance between responsiveness and reliability on the test
system.  YMMV

For operations that have non-blocking built in, e.g. `readv2` with
RWF_NOWAIT set, prefer that to `reactor_promise_nonchalant`.  But for
operations like `poll_add`, apply nonchalance.