operations/poll: make non-blocking request more reliable
Previously, poll had only two timeout modes: blocking for a negative
relative timeout and a conversion to an absolute timeout for everything
else, including a timeout of zero.  However, a relative timeout of zero that
was intended to be non-blocking could spuriously fail since the linked
timeout was always in the past, exacerbated by SQE pressure.

Instead, with non-blocking use nonchalance instead of impatience.