ares_dns: retry epoll_wait on EINTR and allow batch processing
The initial version of this did not allow for `epoll_wait` on more than one
event because sometimes an event would fire but poll would indicate that the
`fd` was not actually ready.  This was the result of flaky non-blocking
behaviour in poll.  Now that poll is more resiliant, allow batch processing.

Note that this does add latency to the a waiter that is performing direct
resolution.  However, once that waiter completes the latency is absorbed by
the resolution fiber.