iou_tls: initial commit
This commit adds iou_tls (a basic TLS client) and TLS support via the
rustls-ffi library [1].

 * iou_rustls_connect: Try to create a rustls_connection on an already
	connected fd from a rustls_client_config_builder.  The builder is
	consumed by this call and should NOT be freed, in contrast to the
	other builders used to set up rustls.

 * iou_rustls_read: Read plaintext bytes from a rustls_connection, reading
	additional ciphertext from fd if necessary.  Returns the number of
	read bytes, 0 on expected EOF, or -EPIPE on expected EOF.  If there
	are not currently consumable plaintext bytes, block on fd.

 * iou_rustls_write: Write plaintext bytes to a rustls_connection, writing
	ciphertext to fd only if necessary.  Returns the number of written
 	bytes, or an error.

 * iou_rustls_flush: Flush all pending plaintext, writing ciphertext to fd.

 * iou_rustls_shutdown: Indicate to peer that we are closing the connection,
	and flush resulting ciphertext.

 * iou_rustls_close: Destroy rustls_connection and close fd.

iou_tls takes a list of HOSTs and HOST/SERVICEs to try to connect to on the
command line with SERVICE defaulting to "echo".  It concurrently does a DNS
lookup for each HOST, and then dials each HOST in command line order, trying
each resolved address until one connects.  If this succeeds, it attempts to
complete a TLS handshake with the host and verifies its name.  On success,
STDIN is connected to TLS egress and TLS ingress is connected to STDOUT.  On
failure, the next HOST is tried until the list is exhausted.

The root store is loaded from /etc/ssl/certs/ca-certificates.crt by default,
but can be overridden in the environment via IOUCONTEXT_CA_BUNDLE.

```bash
	echo -ne 'GET / HTTP/1.0\r\nHost:example.net\r\n\r\n' | \
	./iou_tls 100::1/https example.net/http example.net/https
```

[1] https://github.com/rustls/rustls-ffi