mutex: initial commit
This implements a recursive mutex, although simple mutexes should be
preferable.  When a fiber blocks on enter, it is put into a waitqueue and
woken up when another fiber leaves mutex.  During this time, the reactor is
unable to see the fiber, so reactor_run may return even if not every fiber
is complete.  One should assert that all mutexen are not taken in this case.

iou_mutex_t is ready to use when zero-initialized: `iou_mutex_t my_mutex = { };`

 * iou_mutex_taken: returns true if the mutex has an owner
 * iou_mutex_probe: attempt to enter the mutex, returns true on success
 * iou_mutex_enter: unconditionally enter the mutex, block until possible
 * iou_mutex_owner: returns true if the current fiber owns the mutex
 * iou_mutex_leave: leave the mutex, unblock any waiting fibers

A mutex should not be entered from outside a fiber, nor should any fiber
leave a mutex it does not currently own.