ares_dns: make send synchronous
With the rewrite of the c-ares state machine, it looks like sends that are
asynchronous can lead to reëntrant calls that corrupt channel state.  Light
debugging indicates that when `ares_getaddrinfo` is called with hint
AF_UNSPEC, the second `ares_query_nolock` in `next_dns_lookup` provoke the
issue.  While forcing TCP via `ARES_FLAG_USEVC` also seems to mitigate the
issue, it is simpler to make `send` synchronous for now.