operations/sendto: convert to zero-copy
When issuing a zero-copy send, more than one CQE may come back.  In the case
of a submission error, we will get a single CQE with the error code as a
negative result.  But if the submission succeeds, we will get back the
number of bytes submitted with the `IORING_CQE_F_MORE` set.  In this case,
we expect to recieve a second CQE with the `IORING_CQE_F_NOTIF` and a result
of 0, unless there was an error.  Therefore,

	* when MORE is set: bump the SQE count in anticipation of a future
	  NOTIF, and do not enqueue the `todo`
	* when NOTIF is set: only update the `todo` result if negative

This ensures both that the return value is what we expect --- the number of
bytes written OR an error code --- and that the buffer is now used by write
and available for reuse.