reactor: add synchronize
The `reactor_synchronize` method marks the next operation as a pivot,
preventing the reactor from switching to any other green threads prior to
switching back to the pivot.  In the CQE processor, do NOT invoke the next
`todo` unless it is the `pivot`.  When reserving SQEs, do not yield to the
pending task list, but instead try to drain SQEs/CQEs until there is space
to submit.

It would be nice in the future to only block specific green thread groups.