ares_dns: s/ares_socket_functions/ares_socket_functions_ex/
Update to use the newer socket function object and add supporting functions.

Synchronizes sendto, the callers of which are not reëntrant and will
otherwise corrupt the c-ares channel state.  Poll assertions have a similar
issue in the sendto callback.

Yield to the reactor if no events are reported in `epoll_wait` and remove
the poll assertions, which are failing spuriously.

Remove fcntl/O_NONBLOCK from socket creation, which doesn't seem to make a
difference for c-ares and breaks with a large number of descriptors.