operations/yield: inject NOP based on other pending work
Originally, all async operations would generate an SQE/CQE and yield would
only issue a NOP if there were pending CQEs to process.  When processed,
CQEs would immediately jump to the calling thread.  Now, CQEs enqueue work
into the todo waitqueue and all pending CQEs are processed before attempting
more work.  Thus, yield should issue a NOP if there are any SQEs in flight
OR if there are any todos in the workqueue.  If synchronize is in effect,
also issue the NOP to clear that flag.