mutex: rework with futex, break out recursive version
This commit is ABI breaking; iou_mutex_t is now just an Atomic uint32_t and
the lock is not recursive.  The recursive lock is instead implemented by
iou_mootex_t, which wraps iou_mutex_t and uses `reactor_current` to track
ownership.  References to the running fiber and the reactor's workqueue are
gone, replaced by futex operations.