reactor: add method to get current fiber id
The intention is to use this in a new recursive mutex implementation, but it
might be generally useful to have an stable ID for the currently running
context.  Although it is currently just a pointer to the fiber (or the
reactor, when invoked from outside a thread) the ID should be considered
opaque.