mutex: use helgrind
Add support for mutex tracking.  Unfortunately, helgrind does not understand
fibers, so `iou_mutex_t` is registered as recursive to prevent false
positives.  The non-tracked mutex operations are broken out so that both
mutex and mootex can do their own tracking.  For mootex, owner and depth are
explicitly not tracked since they are part of the lock+unlock protocol.