operations: improve yield
The yield operation does not actually enter the reactor if the reactor has
nothing else to do.  Return whether yield enters the reactor or not.  This
allows busy loops to know whether they should keep spinning or if it is OK
to block.  Allowing them to block means possibly reduced CPU usage.