operations: add flock and friends
Add interfaces for F_OFD_GETLK+F_OFD_SETLK, and helpers to simplify use.
Since io_uring does not currently provide a mechanism to do this, implement
as a busy loop à la the initial version of `iou_waitpid`.  When in blocking
mode, if the reactor has no other work to do, use the F_OFD_SETLKW variant,
which will block the process.  Note that these are per-open-file-descriptor
locks, not per-process locks.

 * iou_fd_getlock: test whether a `struct flock` on `fd` would succeed
 * iou_fd_setlock: attempt to take a `struct flock` on `fd , with optional
	timeout

with helpers

 * iou_fd_lock_append: attempt to take an exclusive lock via `fd` for
	`length` bytes after the end of the file, with optional timeout
 * iou_fd_lock_read: attempt to take a shared lock via `fd` for `length`
	bytes from `start`, with option timeout
 * iou_fd_lock_write: attempt to take an exclusive lock via `fd` for
	`length` bytes from `start`, with optional timeout
 * iou_fd_unlock: remove any lock `fd` holds for `length` bytes from `start`

For all helpers that take `start`, negative values indicate bytes from the
end of the file; otherwise, bytes from the start of the file.  There is no
option to make start relative to the current file since io_uring seems to
encourage explicitly managing offsets rather than relying on the file
position cursor.