eventcount: futex optimization
Use the least significant bit of the ticket to indicate whether there are
any waiters.  If the bit us unset on wait, attempt to set it; on failure,
the ticket is already stale.  If successful, futex wait as usual.  When
waking, unset the bit and only futex wake if it had been set.