iou-cares: use ares_queue_active_queries
The `ares_queue_active_queries` indicates whether there are any queries in
the channel, regardless of whether there are waiters.  Spawn a resolution
thread when there are pending queries, even if there are not yet waiters,
and assert the query queue state when relevant.

Note: resolution will still not begin until after the first waiter waits