slotpool: add try
Add try, which attempts several times to acquire a slot but will not wait on
one becoming available.  On failure, returns the size_t of the pool.  May
fail due to contention even if a slot is available.  Rework `get` in terms
of `try`, and forget trying to avoid the futex.