operations/poll: rework variants
For `in`, `out`, & `hup`: add a set of events that will cause the call to
return false, even if the requested events are present.  Also allow rdhup to
wake up `in`, without causing failure since there may be unread bytes in the
stream even if the peer has closed the socket.