reactor: elide superflous siglongjmps
When the current fiber is the one that issued the next todo, we can just
exit the core without siglongjmp-ing.  This will never happen with ucontext,
for which the enqueuing fiber is different.  It does mean that when a fiber
ends, we must still track it as current until reactor_enter_core either
invokes, jumps to runner, or aborts.