reactor/todos: use multiple queues
Increase the number of `jump_queue_t`s to 3, such that todos in

	`todo[0]` are not blocked on acquiring SQEs,
	`todo[1]` are blocked on acquiring 1 SQE, &
	`todo[2]` are blocked on acquiring 2 SQEs.

In `reactor_enter_core`, dequeue only when queue todo will not immediately
block due to lack of SQEs.  Prioritize todos that need more SQEs, since they
are likely to be timed operations.

As a nice side-effect, immediately run todos that do not require an SQE even
when there are none available.

If any operation ever requires more than 2 SQEs, either more queues must be
added or `reactor_defer` must be modified to accomodate a catch-all.  Since
this function an internal helper, it asserts rather than abort.