iou_h2c: use fixed files
Open the payload file directly as a fixed file.  Accept connections
directly.  Use a POSIX FD in the main thread for listener since we use
shutdown to terminate threads, but register as a fixed file in each thread.
Do not create pipes directly, since we use fcntl to GETPIPE_SZ.  Instead,
create as FDs, fcntl, then transmute to FFs.