output: add polling for short reads, nonblock
If we have a short read or a receive -EAGAIN, poll before any other ops.
The implicit poll op will be in the first sqe/todo.  If poll fails, report
the negative error if no bytes have gone out; otherwise, return the partial
write.  In the case of POLLHUP, this means -EPIPE.  With POLLERR, query
SO_ERROR.